requirements.txt
pandas>=2.1,<3
requests>=2.31,<3
python-dotenv>=1.0,<2
openai>=1.100.0,<2
tenacity>=8.2,<9
